+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Moved
 
 - types: Move `Namespaced` to `types::signaling::NamespacedCommand`
+- types: Move `NamespacedOutgoing` to `types::signaling::NamespacedEvent`
 
 ## [0.2.0] - 2023-03-13